Add files via upload
Started JS Platformer. added in a physicsBody, which should be a part class of all entities that have physics properties. Add in a rect class that mainly draws rectangles. Add in a vec2D so vecotr math can be used for position and movements etc... added in a ground that resizes. input detection. Movement with vel. oop.